refactor(service-bootstrap): inline @mbe/feature-flags, delete one-importer package
The feature-flags package had exactly one importer (service-bootstrap) and
wrapped ~50 lines of header-parse + percentage-rollout logic behind a
plugin-factory boundary that pretended to be a seam. Inline the feature-context
evaluation into service-bootstrap and delete the package. Re-extract if/when a
second flags backend (e.g. PostHog) actually arrives.

- Move header parsing, hashCode hashing, and percentage rollout into
  packages/service-bootstrap/src/feature-flags.ts; request.features decoration
  behavior is byte-for-byte identical (same header name, same hash, same
  plugin name).
- Port both feature-flags test files into feature-flags.test.ts (19 tests).
- Export createFeatureContext/createFeatureFlagsPlugin/FEATURE_FLAGS_HEADER and
  FeatureContext/FeatureFlag types from the service-bootstrap barrel.
- Delete packages/feature-flags entirely; drop the workspace dependency.
- Regenerate pnpm-lock.yaml, infrastructure/worker/dep-graph.json, and
  docs/architecture/dependency-graph.md.

Closes #2085
