+++ infrastructure/worker/dep-graph.json
@@ -66,11 +66,6 @@
       "path": "packages/database"
     },
     {
-      "name": "@mbe/feature-flags",
-      "type": "package",
-      "path": "packages/feature-flags"
-    },
-    {
       "name": "@mbe/jobs",
       "type": "package",
       "path": "packages/jobs"
@@ -418,11 +413,6 @@
       "type": "devDependency"
     },
     {
-      "from": "@mbe/feature-flags",
-      "to": "@mbe/config",
-      "type": "devDependency"
-    },
-    {
       "from": "@mbe/jobs",
       "to": "@mbe/config",
       "type": "devDependency"
@@ -490,11 +480,6 @@
     {
       "from": "@mbe/service-bootstrap",
       "to": "@mbe/database",
-      "type": "dependency"
-    },
-    {
-      "from": "@mbe/service-bootstrap",
-      "to": "@mbe/feature-flags",
       "type": "dependency"
     },
     {
