--- packages/feature-flags/llms.txt
@@ -1,25 +1,0 @@
-<codebase path="packages/feature-flags">
-  <section priority="medium" role="src">
-  <file path="src/index.ts">
-    <item priority="low">
-      interface FeatureFlag {
-        enabled: boolean;
-        percentage: number;
-      }
-    </item>
-    <item priority="low">
-      interface FeatureContext {
-        
-      }
-    </item>
-  </file>
-  <file path="src/plugin.ts">
-    <item priority="high">
-      export const FEATURE_FLAGS_HEADER: unknown;
-    </item>
-    <item priority="high">
-      export function createFeatureFlagsPlugin(): FastifyPluginAsync { /* ... */ }
-    </item>
-  </file>
-  </section>
-</codebase>
