--- packages/feature-flags/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@mbe/feature-flags",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "production": "./dist/index.js",
-      "default": "./src/index.ts"
-    }
-  },
-  "scripts": {
-    "build": "tsc",
-    "lint": "eslint src/",
-    "typecheck": "tsc --noEmit",
-    "test": "vitest run",
-    "test:coverage": "vitest run --coverage"
-  },
-  "devDependencies": {
-    "@mbe/config": "workspace:*",
-    "@types/node": "catalog:",
-    "typescript": "catalog:",
-    "vitest": "catalog:"
-  },
-  "dependencies": {
-    "fastify": "catalog:",
-    "fastify-plugin": "^6.0.0"
-  },
-  "peerDependencies": {
-    "fastify": "catalog:"
-  },
-  "peerDependenciesMeta": {
-    "fastify": {
-      "optional": true
-    }
-  }
-}
