--- packages/feature-flags/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@mbe/config/typescript/node",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "include": ["src"]
-}
