--- packages/feature-flags/vitest.config.ts
@@ -1,21 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: "node",
-    include: ["src/**/*.test.ts"],
-    coverage: {
-      provider: "v8",
-      reporter: ["text", "json", "html"],
-      include: ["src/**/*.ts"],
-      exclude: ["src/**/*.test.ts"],
-      thresholds: {
-        lines: 90,
-        branches: 80,
-        functions: 90,
-        statements: 90,
-      },
-    },
-  },
-});
