+++ packages/service-bootstrap/src/index.ts
@@ -11,3 +11,9 @@ export { createLatencyTracker, checkAuth0 } from "./health.js";
 export type { LatencyTracker, LatencyAnomalyResult, Auth0CheckResult } from "./health.js";
 export { registerHealthRoutes } from "./health-routes.js";
 export type { HealthRoutesOptions, HealthRouteConfig } from "./health-routes.js";
+export {
+  createFeatureContext,
+  createFeatureFlagsPlugin,
+  FEATURE_FLAGS_HEADER,
+} from "./feature-flags.js";
+export type { FeatureContext, FeatureFlag } from "./feature-flags.js";
